backend/AnkiAICardCreationToolboxBackend/pyproject.toml
[project]
name = "ankiaicardcreationtoolboxbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.131.0",
    #"langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    #"langgraph>=1.0.9",
    "pip>=26.0.1",
    "deepagents>=0.4.3",
    "tavily-python>=0.7.21", # required by langgraph
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "httpx>=0.28.1", # required for fastapi testing
]
website_text_extraction = [
    "trafilatura>=2.0.0",
]


[project.scripts]
ankiaicardcreationtoolboxbackend = "ankiaicardcreationtoolboxbackend:main"
create_knowledge_base = "ankiaicardcreationtoolboxbackend:invoke_create_knowledge_base"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"


[tool.uv.workspace]
members = [
    "ankiaicardcreationtoolboxbackend",
]


